Openlms 404 stable v2

+++ lib.php
@@ -177,20 +177,22 @@ function tiny_cursive_upload_multipart_record($filerecord, $filenamewithfullpath
             $filetosend = new CURLFILE($filenamewithfullpath);
         } else {
             // Save base64 decoded content to a temporary JSON file.
-            $tempfilepath = tempnam(sys_get_temp_dir(), 'upload');
-            $filecontent = base64_decode($filerecord->content);
-            $jsoncontent = json_decode($filecontent, true);
+            if (isset($filerecord->content)) {
+                $tempfilepath = tempnam(sys_get_temp_dir(), 'upload');
+                $filecontent = base64_decode($filerecord->content);
+                $jsoncontent = json_decode($filecontent, true);
 
-            if (json_last_error() !== JSON_ERROR_NONE) {
-                throw new Exception("Invalid JSON content in file.");
-            }
-            file_put_contents($tempfilepath, json_encode($jsoncontent));
-            $filetosend = new CURLFILE($tempfilepath, 'application/json', 'uploaded.json');
+                if (json_last_error() !== JSON_ERROR_NONE) {
+                    throw new Exception("Invalid JSON content in file.");
+                }
+                file_put_contents($tempfilepath, json_encode($jsoncontent));
+                $filetosend = new CURLFILE($tempfilepath, 'application/json', 'uploaded.json');
 
-            // Ensure the temporary file does not exceed the size limit.
-            if (filesize($tempfilepath) > 16 * 1024 * 1024) {
-                unlink($tempfilepath);
-                throw new Exception("File exceeds the 16MB size limit.");
+                // Ensure the temporary file does not exceed the size limit.
+                if (filesize($tempfilepath) > 16 * 1024 * 1024) {
+                    unlink($tempfilepath);
+                    throw new Exception("File exceeds the 16MB size limit.");
+                }
             }
         }
 
@@ -288,16 +290,15 @@ function tiny_cursive_get_user_essay_quiz_responses($userid, $courseid, $resourc
                JOIN {quiz} qz ON qa.quiz = qz.id
                JOIN {question} q ON qna.questionid = q.id
                JOIN {course_modules} cm ON qz.id = cm.instance AND cm.module = (
-                   SELECT id FROM {modules} WHERE name = 'quiz'
-               )
-         WHERE qa.userid = :userid
-               AND qz.course = :courseid
-               AND qa.id = :resourceid
-               AND cm.id = :cmid
-               AND q.id = :questionid
-               AND q.qtype = 'essay'
-               AND qas.state = 'complete'
-         ORDER BY qa.attempt, qna.id, qas.sequencenumber";
+                     SELECT id FROM {modules} WHERE name = 'quiz')
+              WHERE qa.userid = :userid
+                    AND qz.course = :courseid
+                    AND qa.id = :resourceid
+                    AND cm.id = :cmid
+                    AND q.id = :questionid
+                    AND q.qtype = 'essay'
+                    AND qas.state = 'complete'
+                    ORDER BY qa.attempt, qna.id, qas.sequencenumber";
 
     $result = $DB->get_record_sql(
         $sql,
@@ -327,16 +328,15 @@ function tiny_cursive_get_user_onlinetext_assignments($userid, $courseid, $modul
     global $DB;
 
     $sql = "SELECT cm.instance as assignmentid, ontext.onlinetext, :modulename AS modulename
-          FROM {assign_submission} asub
-               JOIN {assign} a ON asub.assignment = a.id
-               JOIN {assignsubmission_onlinetext} ontext ON asub.id = ontext.submission
-               JOIN {course_modules} cm ON a.id = cm.instance AND cm.module = (
-                   SELECT id FROM {modules} WHERE name = 'assign'
-               )
-         WHERE asub.userid = :userid
-               AND a.course = :courseid
-               AND asub.status = 'submitted'
-               AND cm.id = :cmid";
+              FROM {assign_submission} asub
+              JOIN {assign} a ON asub.assignment = a.id
+              JOIN {assignsubmission_onlinetext} ontext ON asub.id = ontext.submission
+              JOIN {course_modules} cm ON a.id = cm.instance AND cm.module = (
+                    SELECT id FROM {modules} WHERE name = 'assign')
+             WHERE asub.userid = :userid
+                    AND a.course = :courseid
+                    AND asub.status = 'submitted'
+                    AND cm.id = :cmid";
 
     $result =
         $DB->get_record_sql($sql, ['userid' => $userid, 'courseid' => $courseid, 'modulename' => $modulename, 'cmid' => $cmid]);
@@ -355,12 +355,12 @@ function tiny_cursive_get_user_forum_posts($userid, $courseid, $resourceid) {
     global $DB;
 
     $sql = "SELECT fp.id AS postid, fp.subject, fp.message
-                  FROM {forum_posts} fp
-                       JOIN {forum_discussions} fd ON fp.discussion = fd.id
-                       JOIN {forum} f ON fd.forum = f.id
-                 WHERE fp.userid = :userid
-                       AND fd.course = :courseid
-                       AND fp.id = :resourceid";
+              FROM {forum_posts} fp
+              JOIN {forum_discussions} fd ON fp.discussion = fd.id
+              JOIN {forum} f ON fd.forum = f.id
+             WHERE fp.userid = :userid
+                   AND fd.course = :courseid
+                   AND fp.id = :resourceid";
 
     $result = $DB->get_record_sql($sql, ['userid' => $userid, 'courseid' => $courseid, 'resourceid' => $resourceid]);
     return $result->message;
